Added rpc command:
Disabled direct write access to database table by web interface. This is now handled
by the file_distribution_rpc command which includes a privilege check.

Users on PCs which are online now gets a notification about the file
distribution.

If file distribution is enabled a shortcut to access the files is
created on the desktop of the hosts.